[rb] make OSSBridge and W3CBridge subclasses of Bridge  - REFACTOR
