+++ rb/lib/selenium/webdriver/chrome/bridge.rb
@@ -21,7 +21,7 @@ module Selenium
   module WebDriver
     module Chrome
       # @api private
-      class Bridge < Remote::Bridge
+      class Bridge < Remote::OSSBridge
         def driver_extensions
           [DriverExtensions::TakesScreenshot,
            DriverExtensions::HasInputDevices,
