+++ rb/lib/selenium/webdriver/common/driver.rb
@@ -46,15 +46,12 @@ module Selenium
           bridge = case browser
                    when :firefox, :ff
                      if Remote::W3CCapabilities.w3c?(opts)
-                       if opts[:desired_capabilities].is_a? Remote::Capabilities
-                         opts[:desired_capabilities] = Remote::W3CCapabilities.new(opts[:desired_capabilities].send(:capabilities))
-                       end
                        Firefox::W3CBridge.new(opts)
                      else
                        Firefox::Bridge.new(opts)
                      end
                    when :remote
-                     Remote::Bridge.new(opts)
+                     Remote::OSSBridge.new(opts)
                    when :ie, :internet_explorer
                      IE::Bridge.new(opts)
                    when :chrome
