+++ rb/lib/selenium/webdriver/remote.rb
@@ -22,6 +22,7 @@ require 'uri'
 require 'selenium/webdriver/remote/capabilities'
 require 'selenium/webdriver/remote/w3c_capabilities'
 require 'selenium/webdriver/remote/bridge'
+require 'selenium/webdriver/remote/oss_bridge'
 require 'selenium/webdriver/remote/w3c_bridge'
 require 'selenium/webdriver/remote/server_error'
 require 'selenium/webdriver/remote/response'
